+++ eslint.config.mjs
@@ -40,7 +40,6 @@ export default [
       '@stylistic/eol-last': 'error',
       '@stylistic/function-call-argument-newline': ['error', 'consistent'],
       '@stylistic/function-call-spacing': ['error', 'never'],
-      '@stylistic/func-call-spacing': 'error',
       '@stylistic/function-paren-newline': 'off',
       '@stylistic/generator-star-spacing': 'error',
       '@stylistic/implicit-arrow-linebreak': ['error', 'beside'],