+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Version 6.6.0 (Work in Progress)
 
+#### Enhancements
+
+- The Preferences policy does not allow the use of all Firefox options. There has already been a validation to ensure
+  that no option name is used that is not allowed to be used or does not start with a certain prefix. However, there are
+  also a few options that start with a permitted prefix but are still not allowed. These options are now also recognized
+  as incorrect by the validation, see [#309](https://github.com/cadeyrn/enterprise-policy-generator/issues/309)
+
 #### Dependencies
 
 - Updated eslint from version 9.28.0 to 9.29.0