+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 #### Dependencies
 
 - Updated eslint from version 9.28.0 to 9.29.0
-- Updated eslint-plugin-jsdoc from version 50.8.0 to 51.0.1
+- Updated eslint-plugin-jsdoc from version 50.8.0 to 51.0.3
 
 ---