+++ CHANGELOG.md
@@ -34,7 +34,7 @@
 #### Dependencies
 
 - Updated eslint from version 9.28.0 to 9.29.0
-- Updated eslint-plugin-jsdoc from version 50.8.0 to 51.0.3
+- Updated eslint-plugin-jsdoc from version 50.8.0 to 51.2.1
 - Updated stylelint from version 16.20.0 to 16.21.0
 - Updated web-ext from version 8.7.1 to 8.8.0