+++ CHANGELOG.md
@@ -37,7 +37,7 @@
 - Updated eslint/json from version 0.12.0 to 0.13.1
 - Updated eslint-plugin-jsdoc from version 50.8.0 to 51.3.3
 - Updated stylelint from version 16.20.0 to 16.21.1
-- Updated stylistic/eslint-plugin from version 4.4.1 to 5.1.0
+- Updated stylistic/eslint-plugin from version 4.4.1 to 5.2.2
 - Updated web-ext from version 8.7.1 to 8.8.0
 
 ---