+++ CHANGELOG.md
@@ -38,7 +38,7 @@
 - Updated eslint-plugin-jsdoc from version 50.8.0 to 52.0.2
 - Updated stylelint from version 16.20.0 to 16.23.0 and updated configuration
 - Updated stylistic/eslint-plugin from version 4.4.1 to 5.2.2
-- Updated web-ext from version 8.7.1 to 8.8.0
+- Updated web-ext from version 8.7.1 to 8.9.0
 
 ---