+++ CHANGELOG.md
@@ -37,7 +37,7 @@
 
 #### Dependencies
 
-- Updated eslint from version 9.28.0 to 9.32.0
+- Updated eslint from version 9.28.0 to 9.34.0
 - Updated eslint/json from version 0.12.0 to 0.13.2
 - Updated eslint-plugin-jsdoc from version 50.8.0 to 54.1.1
 - Updated stylelint from version 16.20.0 to 16.23.0 and updated configuration